state of branch correspondes to parallel version

+++ src/BMP.h
@@ -52,7 +52,7 @@ struct BMP {
     void clockwiseRotate();
     void CounterClockwiseRotate();
 
-    void GaussFiltr(float kernelSize);
+    void GaussFiltr(int kernelSize);
 
 
     BMP(const char *fname) {
